# openapi.json
{
  "components": {
    "schemas": {
      "AccountingMethod": {
        "description": "GHG Protocol Scope 2 accounting methods.",
        "enum": [
          "location_based",
          "market_based"
        ],
        "title": "AccountingMethod",
        "type": "string"
      },
...
      "Body_upload_csv_api_v1_compliance_usage_upload_csv_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_csv_api_v1_compliance_usage_upload_csv_post",
        "type": "object"
      },
      "CalculateRequest": {
        "properties": {
          "method": {
            "$ref": "#/components/schemas/AccountingMethod",
            "default": "location_based"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          }
        },
        "required": [
          "org_id"
        ],
        "title": "CalculateRequest",
        "type": "object"
      },
      "CalculationResponse": {
        "properties": {
          "calculations_count": {
            "title": "Calculations Count",
            "type": "integer"
          },
          "data_sources_used": {
            "items": {
              "type": "string"
            },
            "title": "Data Sources Used",
            "type": "array"
          },
          "scope2_kgco2e": {
            "title": "Scope2 Kgco2E",
            "type": "number"
          },
          "scope3_kgco2e": {
            "title": "Scope3 Kgco2E",
            "type": "number"
          },
          "total_emissions_kgco2e": {
            "title": "Total Emissions Kgco2E",
            "type": "number"
          }
        },
        "required": [
          "calculations_count",
          "total_emissions_kgco2e",
          "scope2_kgco2e",
          "scope3_kgco2e",
          "data_sources_used"
        ],
        "title": "CalculationResponse",
        "type": "object"
      },
...
      "ComplianceReportSummary": {
        "description": "Lightweight report listing for API responses.",
        "properties": {
          "carbon_saved_percentage": {
            "title": "Carbon Saved Percentage",
            "type": "number"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "report_name": {
            "title": "Report Name",
            "type": "string"
          },
          "total_energy_kwh": {
            "title": "Total Energy Kwh",
            "type": "number"
          },
          "total_kgco2e": {
            "title": "Total Kgco2E",
            "type": "number"
          }
        },
        "required": [
          "id",
          "report_name",
          "period_start",
          "period_end",
          "generated_at",
          "total_kgco2e",
          "total_energy_kwh",
          "carbon_saved_percentage"
        ],
        "title": "ComplianceReportSummary",
        "type": "object"
      },
...
      "EmissionsCalculation": {
        "description": "A single emissions calculation with full audit trail.",
        "properties": {
          "calculated_at": {
            "format": "date-time",
            "title": "Calculated At",
            "type": "string"
          },
          "emission_factor_gco2_kwh": {
            "title": "Emission Factor Gco2 Kwh",
            "type": "number"
          },
          "emission_factor_quality": {
            "title": "Emission Factor Quality",
            "type": "string"
          },
          "emission_factor_source": {
            "title": "Emission Factor Source",
            "type": "string"
          },
          "emissions_kgco2e": {
            "description": "Total emissions in kg CO2 equivalent",
            "minimum": 0.0,
            "title": "Emissions Kgco2E",
            "type": "number"
          },
          "energy_kwh": {
            "title": "Energy Kwh",
            "type": "number"
          },
          "grid_zone": {
            "title": "Grid Zone",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/AccountingMethod"
          },
          "methodology_version": {
            "default": "ghg_protocol_2024",
            "title": "Methodology Version",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "pue": {
            "default": 1.0,
            "description": "Data center PUE applied",
            "minimum": 1.0,
            "title": "Pue",
            "type": "number"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "renewable_percentage": {
            "title": "Renewable Percentage",
            "type": "number"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/EmissionScope"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "usage_quantity": {
            "title": "Usage Quantity",
            "type": "number"
          },
          "usage_unit": {
            "title": "Usage Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "scope",
          "method",
          "provider",
          "region",
          "grid_zone",
          "service",
          "resource_type",
          "usage_quantity",
          "usage_unit",
          "energy_kwh",
          "emission_factor_gco2_kwh",
          "emission_factor_source",
          "emission_factor_quality",
          "emissions_kgco2e",
          "renewable_percentage",
          "period_start",
          "period_end",
          "calculated_at"
        ],
        "title": "EmissionsCalculation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "carbonlens__compliance__routes__GenerateReportRequest": {
        "properties": {
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "org_name": {
            "title": "Org Name",
            "type": "string"
          },
          "period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "report_name": {
            "default": "",
            "title": "Report Name",
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "org_name"
        ],
        "title": "GenerateReportRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A free, open API for the real-time carbon intensity of cloud regions. No account or API key required; usage is rate-limited so it stays responsive for everyone.\n\n## Carbon Data API\n- **8 live grid-operator integrations** (UK, EIA, OpenElectricity/AEMO, IESO/AESO, Taipower, GridStatus, ENTSO-E, Electricity Maps) plus labeled heuristic and weather-based estimates, cascading\n- **Grid carbon intensity** for 75+ cloud regions; every response is tagged with its `source`\n- **Batch queries** for multiple regions in a single call\n\n## Compliance Reporting\n- **GHG-Protocol-structured** Scope 2 (location-based) + Scope 3 Cat 1 reporting\n- Drafts aimed at **CSRD / SEC Climate / SB 253** workflows (not an assured report)\n- **EU Taxonomy** eligibility flag (simplified screening)\n\n## Green SLA Monitoring (Beta)\n- **Define carbon targets** \u2014 max gCO2/kWh, min renewable %\n- **On-demand + background checks** (in-memory state)\n- **Attestation-style summary reports**\n\n## Carbon-Aware Scheduling\n- **Find optimal time windows** for batch jobs, CI/CD, ML training\n- **Multi-region evaluation** across AWS, GCP, Azure\n- **Three strategies** \u2014 lowest carbon, highest renewable, balanced\n- Returns a recommendation \u2014 it doesn't run or defer your workload itself",
    "title": "Carbon Lens",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/compliance/calculate": {
      "post": {
        "description": "Calculate emissions from ingested usage data for an organization.",
        "operationId": "calculate_emissions_api_v1_compliance_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Calculate Emissions",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/calculations": {
      "get": {
        "description": "List emissions calculations for an organization.",
        "operationId": "list_calculations_api_v1_compliance_calculations_get",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmissionsCalculation"
                  },
                  "title": "Response List Calculations Api V1 Compliance Calculations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Calculations",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/reports": {
      "get": {
        "description": "List all compliance reports for an organization.",
        "operationId": "list_reports_api_v1_compliance_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComplianceReportSummary"
                  },
                  "title": "Response List Reports Api V1 Compliance Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Reports",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/reports/generate": {
      "post": {
        "description": "Generate a CSRD-aligned compliance report from calculated emissions.",
        "operationId": "generate_report_api_v1_compliance_reports_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/carbonlens__compliance__routes__GenerateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Generate Report",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/reports/{report_id}": {
      "get": {
        "description": "Get a specific compliance report by ID.",
        "operationId": "get_report_api_v1_compliance_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Report",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/reports/{report_id}/export": {
      "get": {
        "description": "Export a compliance report as JSON or CSV.",
        "operationId": "export_report_api_v1_compliance_reports__report_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "pattern": "^(json|csv)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Export Report",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/usage": {
      "get": {
        "description": "List ingested usage records for an organization.",
        "operationId": "list_usage_api_v1_compliance_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CloudUsageRecord"
                  },
                  "title": "Response List Usage Api V1 Compliance Usage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Usage",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/usage/ingest": {
      "post": {
        "description": "Ingest cloud usage data from a provider or mock data for demo.",
        "operationId": "ingest_usage_api_v1_compliance_usage_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageIngestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageIngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Ingest Usage",
        "tags": [
          "Compliance"
        ]
      }
    },
    "/api/v1/compliance/usage/upload-csv": {
      "post": {
        "description": "Upload a CSV file with cloud usage data.",
        "operationId": "upload_csv_api_v1_compliance_usage_upload_csv_post",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_csv_api_v1_compliance_usage_upload_csv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageIngestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Upload Csv",
        "tags": [
          "Compliance"
        ]
      }
    },
...
  },
...
}